Matlab code by Gorish Aggarwal
A Brain Computer Interface (BCI) is a direct communication link between
the brain and an
external device. The EEG signal captured through this device can be used
to stimulate the motor neurons
through a Functional Electrical Stimulator, thus allowing the brain to
control the muscles externally. This
requires the processing of the EEG signal, to extract features
corresponding to the task imaged by the brain.
This report gives the description of the internship project undertaken
to develop a mechanism such as
above, which can collect EEG data, process the acquired data, extract
relevant information and use that
information to trigger an FES device for stimulating motor muscles. The
developed system can be used to
develop a low cost, clinically valid home-based tool to monitor
post-stroke neuro-rehabilitation device for
use both in rural and urban setting.